requirements.txt
Django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-environ>=0.11
django-filter>=24.3
django-cors-headers>=4.4
whitenoise>=6.6
gunicorn>=21.2
psycopg[binary]>=3.1